requirements.txt
openai>=1.0.0
playwright==1.41.0
python-dotenv==1.0.0
pydantic==2.6.1
beautifulsoup4==4.12.3
requests==2.31.0
loguru==0.7.2
httpx>=0.24.0

